# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/member_histories:
    get:
      tags:
        - MemberHistories
      summary: List MemberHistories
      description: Get a paginated list of MemberHistories with optional filtering and sorting
      operationId: listMemberHistories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberHistoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MemberHistories
      summary: Create MemberHistory
      description: Create a new MemberHistory
      operationId: createMemberHistory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberHistoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/operator_expertises:
    get:
      tags:
        - OperatorExpertises
      summary: List OperatorExpertises
      description: Get a paginated list of OperatorExpertises with optional filtering and sorting
      operationId: listOperatorExpertises
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorExpertiseList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - OperatorExpertises
      summary: Create OperatorExpertise
      description: Create a new OperatorExpertise
      operationId: createOperatorExpertise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperatorExpertiseRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorExpertise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    MemberHistory:
      type: object
      required:
        - id
        - session_id
        - persona
        - joined_at
        - message_count
        - expertise_names
        - metadata
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        persona:
          type: string
        operator_user_id:
          nullable: true
          type: string
          format: uuid
        visitor_key:
          nullable: true
          type: string
          maxLength: 120
        chatbot_script_id:
          nullable: true
          type: string
          format: uuid
        joined_at:
          type: string
          format: date-time
        left_at:
          nullable: true
          type: string
          format: date-time
        message_count:
          type: integer
          format: int32
        response_time_secs:
          nullable: true
          type: integer
          format: int32
        expertise_names:
          type: array
        metadata:
          type: object
...
    CreateMemberHistoryRequest:
      type: object
      required:
        - session_id
        - persona
        - joined_at
        - message_count
        - expertise_names
        - metadata
      properties:
        session_id:
          type: string
          format: uuid
        persona:
          type: string
        operator_user_id:
          nullable: true
          type: string
          format: uuid
        visitor_key:
          nullable: true
          type: string
          maxLength: 120
        chatbot_script_id:
          nullable: true
          type: string
          format: uuid
        joined_at:
          type: string
          format: date-time
        left_at:
          nullable: true
          type: string
          format: date-time
        message_count:
          type: integer
          format: int32
        response_time_secs:
          nullable: true
          type: integer
          format: int32
        expertise_names:
          type: array
        metadata:
          type: object
...
    MemberHistoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemberHistory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    OperatorExpertise:
      type: object
      required:
        - id
        - operator_profile_id
        - expertise_tag_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        operator_profile_id:
          type: string
          format: uuid
        expertise_tag_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateOperatorExpertiseRequest:
      type: object
      required:
        - operator_profile_id
        - expertise_tag_id
        - metadata
      properties:
        operator_profile_id:
          type: string
          format: uuid
        expertise_tag_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    OperatorExpertiseList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OperatorExpertise'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []